backend/requirements.txt
fastapi
uvicorn
sqlalchemy
databases[sqlite]
pydantic
python-multipart
python-dotenv
passlib[bcrypt]
pyjwt
email-validator
